✨ Started SSF work
Started to work on adding SSF and did some more documentation, played around with psf but realized we need more memory management functions than we have. will work on them later but for now putting aside for simpler SSF just to get proof of concept working. Did some more documentation work as well and modified the Makefiles a bit but overall not too bad. Leaving some work for tomorrow but it is 23:40 now.